Recombine NotablePerson component
