+++ src/app/pages/NotablePerson/NotablePerson.module.scss
